Added a removal routine for the socket

+++ shtdlib.sh
@@ -706,7 +706,7 @@ function priv_esc_with_env {
 function get_custom_ssh_auth_agent {
     docker_ssh_auth_socket_path="${1:-${HOME}/docker-ssh-agent}"
     docker_ssh_auth_pid_file="${2:-${HOME}/.docker-ssh-agent.pid}"
-    if [ -S "${docker_ssh_auth_socket_path}" ] && pgrep --pidfile ${docker_ssh_auth_pid_file} 2> /dev/null ; then
+    if [ -S "${docker_ssh_auth_socket_path}" ] && pgrep -F ${docker_ssh_auth_pid_file} 2> /dev/null ; then
         color_echo cyan "Found docker specific ssh-agent with socket: ${docker_ssh_auth_socket_path}"
         export SSH_AUTH_SOCK="${docker_ssh_auth_socket_path}"
         if [ -f "${docker_ssh_auth_pid_file}" ] ; then
@@ -716,8 +716,13 @@ function get_custom_ssh_auth_agent {
     else
         color_echo cyan "Creating docker specific ssh-agent with socket: ${docker_ssh_auth_socket_path}"
         assert whichs ssh-agent
-        eval $(ssh-agent -a ${docker_ssh_auth_socket_path})
-        echo "${SSH_AGENT_PID}" > "${docker_ssh_auth_pid_file}"
+        if rm -f ${docker_ssh_auth_socket_path} ; then
+            eval $(ssh-agent -a ${docker_ssh_auth_socket_path})
+            echo "${SSH_AGENT_PID}" > "${docker_ssh_auth_pid_file}"
+        else
+            color_echo red "Unable to reset/create named socket ${docker_ssh_auth_socket_path}, please verify path and permissions"
+            return 1
+        fi
     fi
 
     color_echo cyan "Checking ssh-agent key status"
